fix: implement toTitleCase in designs name
