+++ apps/frontend/src/lib/design/index.ts
@@ -1,3 +1,2 @@
 export * from './image-rules';
 export * from './validate-image';
-export * from './title-case';
